moved accordion.tsx to components
